Refactor namespace name and update DialogueNode to have a behaviour instance
