+++ Assets/Scripts/Editor/BehaviourTreeDrawer.cs
@@ -3,11 +3,18 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEditor;
 
-namespace BehaviourTree
+namespace Behaviours
 {
     [CustomPropertyDrawer(typeof(BehaviourTree))]
     public class BehaviourTreeDrawer : PropertyDrawer
     {
+        /// ~~~  TODO  ~~~
+        /// behaviour nodes need to be able to read the instance properties from their behaviour instance
+        /// do we pass these in as a second reference to the Tick() method? this is pretty decent
+        /// another issue... the serializedObject may not reference the IBehaviourInstance directly (see dialogue...)
+        /// does this remove the usefulness of IBehaviourInstance entirely? are we better off making it entirely case-based?
+        /// or do we still need IBehaviourInstance for handling behaviour node logic anyway?
+
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
             SerializedObject owner = property.serializedObject;
@@ -57,8 +64,8 @@ namespace BehaviourTree
         public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
         {
             IBehaviourInstance behaviourInstance = property.serializedObject.targetObject as IBehaviourInstance;
-            int instancePropsHeight = behaviourInstance.GetInstanceProperties().Length * 20 + 24;
-            return base.GetPropertyHeight(property, label) + (behaviourInstance.GetBehaviourTree() != null ? instancePropsHeight : 0);
+            int instancePropsHeight = behaviourInstance.GetBehaviourTree() != null ? behaviourInstance.GetInstanceProperties().Length * 20 + 24 : 0;
+            return base.GetPropertyHeight(property, label) + instancePropsHeight;
         }
 
         private bool DrawInstanceProperty(Rect position, BehaviourInstanceProperty instanceProperty)
