+++ Assets/Scripts/Tools/Behaviour Tree/BehaviourInstanceProperty.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace BehaviourTree
+namespace Behaviours
 {
     [System.Serializable]
     public class BehaviourInstanceProperty
