+++ Assets/Scripts/Tools/Behaviour Tree/BehaviourTreeNodeCreator.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace BehaviourTree
+namespace Behaviours
 {
     public enum BehaviourTreeNodeType
     {
