+++ Assets/Scripts/Tools/Behaviour Tree/Data/DynamicMoveType.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace BehaviourTree
+namespace Behaviours
 {
     public abstract class DynamicMoveType : MoveType
     {
