+++ Assets/Scripts/Tools/Behaviour Tree/Data/MoveType.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace BehaviourTree
+namespace Behaviours
 {
     public abstract class MoveType : ScriptableObject
     {
