+++ Assets/Scripts/Tools/Behaviour Tree/Nodes/GetVectorAxisNode.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace BehaviourTree
+namespace Behaviours
 {
     public enum VectorAxis { X, Y }
