+++ Assets/Scripts/Tools/Behaviour Tree/Nodes/IsEnemyNode.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using RPG;
 
-namespace BehaviourTree
+namespace Behaviours
 {
     // TODO: turn this into HasFactionNode later!
     public class IsEnemyNode : IBehaviourTreeNode
