+++ Assets/Scripts/Tools/Behaviour Tree/Nodes/SequenceNode.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace BehaviourTree
+namespace Behaviours
 {
     public class SequenceNode : IBehaviourTreeNode
     {
