+++ Assets/Scripts/Tools/Behaviour Tree/Nodes/SubTreeNode.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace BehaviourTree
+namespace Behaviours
 {
     public class SubTreeNode : IBehaviourTreeNode
     {
